Minor fix - incompatible type declaration

+++ aes/tiny_aes/aes.h
@@ -41,6 +41,6 @@ struct AES_ctx
 
 void AES_init_ctx(struct AES_ctx* ctx, const uint8_t* key);
 void AES_encrypt(struct AES_ctx* ctx, uint8_t key[], uint8_t in[], uint8_t out[]);
-int * AES_decrypt(struct AES_ctx* ctx, uint8_t key[], uint8_t out[], uint8_t in[]);
+uint8_t * AES_decrypt(struct AES_ctx* ctx, uint8_t key[], uint8_t out[], uint8_t in[]);
 
 #endif // _AES_H_
